Fix direction of friend's printers ownership.
We used to be able to print on printers of people we follow on twitter
but that makes no sense: we want only people *we* follow intentionally
to be able to print on our printers.

This implementation is slow-ish but I added a test case so we can replace
the slow query with a N:M relationship in sqlalchemy in the future.